LayoutTests/css3/scroll-snap/scroll-padding-mainframe-paging-expected.txt

PASS paging moved body
PASS paging moved padded body
PASS document.scrollingElement.scrollTop is within 1 of 535.6837606837607
PASS paging moved padded body
PASS document.scrollingElement.scrollTop is within 1 of 535.6837606837607
PASS paging moved padded body
PASS document.scrollingElement.scrollTop is within 2 of 526.3675213675214
PASS successfullyParsed is true

TEST COMPLETE
